pyproject.toml
[project]
name = "test_agent"
version = "0.1.0"
description = "test_agent using crewAI"
authors = [{ name = "sandipan seal", email = "sandipanseal68@gmail.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai-tools[mcp]>=0.59.0",
    "crewai[tools]>=0.152.0,<1.0.0",
    "mcp>=1.13.0",
    "python-dotenv>=1.1.1",
]

[project.scripts]
test_agent = "test_agent.main:run"
run_crew = "test_agent.main:run"
train = "test_agent.main:train"
replay = "test_agent.main:replay"
test = "test_agent.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
